Unify the schema loading flows
